robot_management/requirements.txt
# Core dependencies
paho-mqtt==1.6.1
pyyaml==6.0.1
python-dateutil==2.8.2

# Development and testing
pytest==7.4.0
pytest-mock==3.11.1
pytest-cov==4.1.0

# Logging and error handling
structlog==23.1.0

# Code quality
black==23.7.0
flake8==6.1.0
mypy==1.4.1
